+++ artifactory/commands/huggingface/huggingFaceUpload.go
@@ -11,7 +11,8 @@ import (
 	"time"
 
 	"github.com/jfrog/build-info-go/entities"
-	"github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
+	"github.com/jfrog/jfrog-cli-artifactory/artifactory/utils"
+	coreUtils "github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
 	buildUtils "github.com/jfrog/jfrog-cli-core/v2/common/build"
 	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
@@ -127,7 +128,7 @@ func (hfu *HuggingFaceUpload) CollectArtifactsForBuildInfo() error {
 
 // GetArtifacts returns HuggingFace model/dataset files in JFrog Artifactory
 func (hfu *HuggingFaceUpload) GetArtifacts(buildProperties string) ([]entities.Artifact, error) {
-	serviceManager, err := utils.CreateServiceManager(hfu.serverDetails, -1, 0, false)
+	serviceManager, err := coreUtils.CreateServiceManager(hfu.serverDetails, -1, 0, false)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create services manager: %w", err)
 	}
@@ -146,7 +147,7 @@ func (hfu *HuggingFaceUpload) GetArtifacts(buildProperties string) ([]entities.A
 		hfu.repoId,
 		revisionPattern,
 	)
-	results, err := executeAqlQuery(serviceManager, aqlQuery)
+	results, err := utils.ExecuteAqlQuery(serviceManager, aqlQuery)
 	if err != nil {
 		return nil, fmt.Errorf("failed to search for HuggingFace artifacts: %w", err)
 	}
